pull out parser from UsersConfigAccessStorage
